Use ordered daemon as default
